REFACTOR: Changed skeleton directory
